sd-replacement-server/requirements.txt
asgiref==3.3.4
click==8.0.1
fastapi==0.65.2
h11==0.12.0
MouseInfo==0.1.3
PyAutoGUI==0.9.52
pydantic==1.8.2
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.27
python3-xlib==0.15
PyTweening==1.0.3
PyYAML==5.4.1
starlette==0.14.2
typing-extensions==3.10.0.0
uvicorn==0.14.0
